requirements.txt
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.22
langchain-openai==0.0.5
pypdf==4.0.1
python-dotenv==1.0.1
streamlit==1.31.0
altair==5.2.0
tiktoken==0.5.2
faiss-cpu==1.7.4
